CODE/requirements.txt
# AWS SDK
boto3>=1.34.69
botocore>=1.34.69

# Streamlit Dashboard
streamlit>=1.32.2
plotly>=5.20.0
pandas>=2.2.1

# AI Agent
openai>=1.14.3
langchain>=0.1.13
langchain-openai>=0.1.1
langchain-community>=0.0.29
langgraph>=0.0.37

# Utilities
python-dotenv>=1.0.1
requests>=2.31.0
tabulate>=0.9.0
humanize>=4.9.0
pytz>=2024.1
